% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{classificadorPico}
\alias{classificadorPico}
\title{Cria labels de pico, vale e nivel}
\usage{
classificadorPico(y, lag, threshold, influence)
}
\arguments{
\item{y}{Vector. Serie temporal.}

\item{lag}{Numeric. Tamanho da janela desejada para media movel}

\item{threshold}{Numeric. Numero de desvios padroes.}

\item{influence}{Numeric. Influence = 0 recalcula o valor do threshold.}
}
\value{
list com os seguintes objetos: signals, avgFilter and stdFilter.
}
\description{
Essa funcao classifica a serie temporal em pico, vale e nivel.
}
\examples{
set.seed(412)
df = tibble(serie = rnorm(100, 10, 1))
plot(df$serie, type = 'l')
df$sinal <- sinais::classificadorPico(df$serie, lag = 6, threshold = 2, influence = 0.1)$signals
head(df)
}
\references{
https://stackoverflow.com/questions/22583391/peak-signal-detection-in-realtime-timeseries-data
}
