% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{featuresSinais}
\alias{featuresSinais}
\title{Features de sinais}
\usage{
featuresSinais(df, defasagem = 1, janela = 6)
}

\arguments{
\item{df}{Dataframe}

\item{defasagem}{Numeric. Por default utiliza o lag 1 da coluna de sinais.}

\item{janela}{Numeric. Tamanho da janela rolante. Por default 6.}
}
\value{
Retorna dataframe.
}


\description{
Cria todas as features de sinais para o modelo de classificacao.
}


\examples{
set.seed(412)
df = tibble(serie = rnorm(100, 10, 1))
df$sinal <- sinais::classificadorPico(df$serie, lag = 6, threshold = 2, influence = 0.1)$signals
sinais::featuresSinais(df, janela = 3) %>% View
}
